Move print sorting functions to utils
