+++ server/routes/reporting/standard-report/standard-report-print.js
@@ -1,7 +1,7 @@
 /* eslint-disable strict */
 const moment = require('moment');
 const { generateDocument } = require('../../../lib/reports/single-generator');
-const { tableDataMappers, constructPageHeading } = require('./utils');
+const { tableDataMappers, constructPageHeading, sort } = require('./utils');
 const { bespokeReportTablePrint } = require('../bespoke-report/bespoke-report-print');
 const { snakeToCamel, checkIfArrayEmpty } = require('../../../lib/mod-utils');
 const { reportKeys } = require('./definitions');
@@ -330,61 +330,12 @@ function sortTableData(reportKey, { sortBy, sortDirection }, tableData, reportDa
   }
 }
 
-function sort(sortBy, sortDirection) {
-  return (a, b) => {
-    const [_a, _b] = formatSortableData(a, b, sortBy);
-      
-    if (isNumber(_a) && isNumber(_b)) {
-      return sortDirection === 'descending' ? _b - _a : _a - _b;
-    }
-
-    if (sortDirection === 'descending') {
-      return _b.localeCompare(_a);
-    } else {
-      return _a.localeCompare(_b);
-    }
-  }
-}
-
-function formatSortableData(a, b, sortBy) {
-  let _a = a[snakeToCamel(sortBy)] || '-';
-  let _b = b[snakeToCamel(sortBy)] || '-';
-
-  if (sortBy === 'jurorPostalAddress') {
-    _a = Object.values(a.jurorPostalAddress).join(' ');
-    _b = Object.values(b.jurorPostalAddress).join(' ');
-  }
-
-  if (sortBy === 'jurorReasonableAdjustmentWithMessage') {
-    _a = a.jurorReasonableAdjustmentWithMessage
-      ? Object.values(a.jurorReasonableAdjustmentWithMessage).join(' ') : '-';
-    _b = b.jurorReasonableAdjustmentWithMessage
-      ? Object.values(b.jurorReasonableAdjustmentWithMessage).join(' ') : '-';
-  }
-
-  if (sortBy === 'contactDetails') {
-    _a = a.contactDetails ? Object.values(a.contactDetails).join(' ') : '-';
-    _b = b.contactDetails ? Object.values(b.contactDetails).join(' ') : '-';
-  }
-
-  if (sortBy === 'month') {
-    _a = dateFilter(a.month, 'mmmm yyyy', 'yyyy-MM-DD');
-    _b = dateFilter(b.month, 'mmmm yyyy', 'yyyy-MM-DD');
-  }
-
-  return [_a.toString(), _b.toString()];
-}
-
 function resolveSortBy(sortBy, reportData) {
   if (sortBy) {
     return sortBy;
   }
 
   return reportData.defaultSortColumn;
-}
-
-function isNumber(n) {
-  return !isNaN(parseFloat(n)) && !(moment(n, 'yyyy-MM-DD', true).isValid());
 }
 
 module.exports = {
