+++ server/routes/reporting/standard-report/utils.js
@@ -1,6 +1,7 @@
 /* eslint-disable strict */
-const { dateFilter, capitalizeFully, toSentenceCase, capitalise, toCamelCase } = require('../../../components/filters');
+const { dateFilter, capitalizeFully, toSentenceCase, capitalise, toCamelCase, makeDate } = require('../../../components/filters');
 const moment = require('moment');
+const { snakeToCamel } = require('../../../lib/mod-utils');
 
 const tableDataMappers = {
   String: (data) => isNaN(data) ? capitalizeFully(data) : (data?.toString() || '-' ),
@@ -126,6 +127,61 @@ const buildTableHeaders = (reportType, tableHeadings, query = {}) => {
   return tableHeaders;
 };
 
+function sort(sortBy, sortDirection) {
+  return (a, b) => {
+    const [_a, _b] = formatSortableData(a, b, sortBy);
+      
+    if (isNumber(_a) && isNumber(_b)) {
+      return sortDirection === 'descending' ? _b - _a : _a - _b;
+    }
+
+    if (sortDirection === 'descending') {
+      return _b.localeCompare(_a);
+    } else {
+      return _a.localeCompare(_b);
+    }
+  }
+}
+
+function formatSortableData(a, b, sortBy) {
+  let _a = a[snakeToCamel(sortBy)] || '-';
+  let _b = b[snakeToCamel(sortBy)] || '-';
+
+  if (sortBy === 'jurorPostalAddress') {
+    _a = Object.values(a.jurorPostalAddress).join(' ');
+    _b = Object.values(b.jurorPostalAddress).join(' ');
+  }
+
+  if (sortBy === 'jurorReasonableAdjustmentWithMessage') {
+    _a = a.jurorReasonableAdjustmentWithMessage
+      ? Object.values(a.jurorReasonableAdjustmentWithMessage).join(' ') : '-';
+    _b = b.jurorReasonableAdjustmentWithMessage
+      ? Object.values(b.jurorReasonableAdjustmentWithMessage).join(' ') : '-';
+  }
+
+  if (sortBy === 'contactDetails') {
+    _a = a.contactDetails ? Object.values(a.contactDetails).join(' ') : '-';
+    _b = b.contactDetails ? Object.values(b.contactDetails).join(' ') : '-';
+  }
+
+  if (sortBy === 'month') {
+    _a = dateFilter(a.month, 'mmmm yyyy', 'yyyy-MM-DD');
+    _b = dateFilter(b.month, 'mmmm yyyy', 'yyyy-MM-DD');
+  }
+
+  if (sortBy === 'date' && (Array.isArray(_a) && Array.isArray(_b))) {
+    _a = dateFilter(makeDate(_a), null, 'yyyy-MM-DD');
+    _b = dateFilter(makeDate(_b), null, 'yyyy-MM-DD');
+  }
+
+  return [_a.toString(), _b.toString()];
+}
+
+function isNumber(n) {
+  return !isNaN(parseFloat(n)) && !(moment(n, 'yyyy-MM-DD', true).isValid());
+}
+
 module.exports.tableDataMappers = tableDataMappers;
 module.exports.constructPageHeading = constructPageHeading;
 module.exports.buildTableHeaders = buildTableHeaders;
+module.exports.sort = sort;
